+++ libifstate/__init__.py
@@ -29,7 +29,7 @@ import re
 import json
 import errno
 
-__version__ = "1.5.6"
+__version__ = "1.5.7"
 
 
 class IfState():